need to fix test. animate when slide/drag is cancelled
